[LibOS] Move locking from `shim_internal.h` to a dedicated header
