+++ LibOS/shim/src/bookkeep/shim_handle.c
@@ -11,6 +11,7 @@
 #include "shim_fs.h"
 #include "shim_handle.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_thread.h"
 
 static struct shim_lock handle_mgr_lock;
