+++ LibOS/shim/src/bookkeep/shim_signal.c
@@ -17,6 +17,7 @@
 #include "shim_checkpoint.h"
 #include "shim_handle.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_signal.h"
 #include "shim_table.h"
 #include "shim_thread.h"
