+++ LibOS/shim/src/bookkeep/shim_vma.c
@@ -24,6 +24,7 @@
 #include "shim_fs.h"
 #include "shim_handle.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_tcb.h"
 #include "shim_utils.h"
 #include "shim_vma.h"
