+++ LibOS/shim/src/elf/shim_rtld.c
@@ -20,6 +20,7 @@
 #include "shim_fs.h"
 #include "shim_handle.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_table.h"
 #include "shim_thread.h"
 #include "shim_utils.h"
