+++ LibOS/shim/src/fs/chroot/fs.c
@@ -6,15 +6,16 @@
  */
 
 // FIXME: Sorting these includes causes a bunch of "error: ‘S_IFREG’ undeclared" errors.
-#include "shim_flags_conv.h"
-#include "shim_internal.h"
-#include "shim_thread.h"
-#include "shim_handle.h"
-#include "shim_vma.h"
-#include "shim_fs.h"
-#include "shim_utils.h"
 #include "pal.h"
 #include "pal_error.h"
+#include "shim_flags_conv.h"
+#include "shim_fs.h"
+#include "shim_handle.h"
+#include "shim_internal.h"
+#include "shim_lock.h"
+#include "shim_thread.h"
+#include "shim_utils.h"
+#include "shim_vma.h"
 
 #include <asm/fcntl.h>
 #include <asm/mman.h>
