+++ LibOS/shim/src/fs/shim_dcache.c
@@ -12,6 +12,7 @@
 #include "shim_fs.h"
 #include "shim_handle.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_types.h"
 
 static struct shim_lock dcache_mgr_lock;
