+++ LibOS/shim/src/fs/shim_fs.c
@@ -14,6 +14,7 @@
 #include "shim_checkpoint.h"
 #include "shim_fs.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_utils.h"
 
 struct shim_fs {
