+++ LibOS/shim/src/fs/shim_namei.c
@@ -14,6 +14,7 @@
 #include "shim_fs.h"
 #include "shim_handle.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_thread.h"
 #include "shim_utils.h"
