+++ LibOS/shim/src/fs/socket/fs.c
@@ -18,6 +18,7 @@
 #include "pal_error.h"
 #include "shim_fs.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_thread.h"
 
 static int socket_close(struct shim_handle* hdl) {
