+++ LibOS/shim/src/ipc/shim_ipc.c
@@ -14,6 +14,7 @@
 #include "shim_handle.h"
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_thread.h"
 #include "shim_types.h"
 #include "shim_utils.h"
