+++ LibOS/shim/src/ipc/shim_ipc_child.c
@@ -13,6 +13,7 @@
 #include "shim_handle.h"
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_thread.h"
 #include "shim_utils.h"
