+++ LibOS/shim/src/ipc/shim_ipc_pid.c
@@ -13,6 +13,7 @@
 #include "shim_fs.h"
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_thread.h"
 
 static int thread_add_subrange(struct shim_thread* thread, void* arg) {
