+++ LibOS/shim/src/ipc/shim_ipc_ranges.c
@@ -9,6 +9,7 @@
 
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_utils.h"
 
 #define BITS (sizeof(char) * 8)
