+++ LibOS/shim/src/shim_checkpoint.c
@@ -19,6 +19,7 @@
 #include "shim_handle.h"
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_thread.h"
 #include "shim_utils.h"
 #include "shim_vma.h"
