+++ LibOS/shim/src/shim_init.c
@@ -21,6 +21,7 @@
 #include "shim_handle.h"
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_table.h"
 #include "shim_tcb.h"
 #include "shim_thread.h"
