+++ LibOS/shim/src/shim_malloc.c
@@ -15,6 +15,7 @@
 #include "pal_debug.h"
 #include "shim_checkpoint.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_utils.h"
 #include "shim_vma.h"
