+++ LibOS/shim/src/sys/shim_alarm.c
@@ -8,6 +8,7 @@
 #include <stdint.h>
 
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_signal.h"
 #include "shim_table.h"
 #include "shim_thread.h"
