+++ LibOS/shim/src/sys/shim_brk.c
@@ -12,6 +12,7 @@
 #include "pal.h"
 #include "shim_checkpoint.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_table.h"
 #include "shim_utils.h"
 #include "shim_vma.h"
