+++ LibOS/shim/src/sys/shim_epoll.c
@@ -14,6 +14,7 @@
 #include "shim_fs.h"
 #include "shim_handle.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_table.h"
 #include "shim_thread.h"
