+++ LibOS/shim/src/sys/shim_exit.c
@@ -10,6 +10,7 @@
 #include "shim_handle.h"
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_table.h"
 #include "shim_thread.h"
 #include "shim_utils.h"
