+++ LibOS/shim/src/sys/shim_getpid.c
@@ -16,6 +16,7 @@
 #include "shim_checkpoint.h"
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_table.h"
 #include "shim_thread.h"
