+++ LibOS/shim/src/sys/shim_semget.c
@@ -13,6 +13,7 @@
 #include "shim_handle.h"
 #include "shim_internal.h"
 #include "shim_ipc.h"
+#include "shim_lock.h"
 #include "shim_sysv.h"
 #include "shim_table.h"
 #include "shim_utils.h"
