+++ LibOS/shim/src/sys/shim_wait.c
@@ -13,6 +13,7 @@
 #include "pal.h"
 #include "pal_error.h"
 #include "shim_internal.h"
+#include "shim_lock.h"
 #include "shim_signal.h"
 #include "shim_table.h"
 #include "shim_thread.h"
